Role add and removal (#74)
* fix mute restoring, make inf search fancier

* fix serverinfo

* more cleaning

* fix message finder failing when edit logs where disabled

* more fixing

* move cogs to load to master config

* allow disabling docs

* missing await

* update requirements

* add rowboat migrator

* fix confirmator for twermoji

* don't lower cog names, it only breaks things

* fix blacklist being shown every time something is added/removed

* limit infractions

* fix mute restoring

* Role add and removal

This is a basic version, which should work other than an known bug where using the role ID does not work. I've made it so that it didn't error out.

* Changed up a bit, added more checks

Changed the code so that it can easily be translated, as well adding in a few more checks as I adjusted it. :)