Mass actions rework (#298)
* Add mcleanban and mwarn

* [BREAKING CHANGES] Migrate to Utils

Migrating went well, but it meant that I had to change the self._can_act to await Utils._can_act if we wanted to avoid any attribute errors. So far it looks like everything is working properly as it should be in terms of the intended behavior.

* Forgot about the empty_list

* Add limit (10 people can be warned).

Not complete, needs to be translated.

* Add string for >10 people warning

* Add mwarn duplicate

* Add the duplicate filtering for the remaining of the mcommands.

They seem to be working properly during the testing. So we should be fine.

* check guild permissions instead of channel perms for muting
fixes #294

* rework mass actions

* make code a bit easier to follow

* missed a few things